refactor: split code into separate files for better modularity
