+++ public/assets/images/icons/myForestMapIcons/RegisteredTreePopupIcon.tsx
