--- public/assets/images/icons/myForestV2Icons/ClusterMarker/RegisteredTreeClusterMarkerIcon.tsx
@@ -1,79 +1,0 @@
-import { IconProps } from '../../../../../../src/features/common/types/common';
-const RegisteredTreeClusterMarkerIcon = ({ width }: IconProps) => {
-  return (
-    <svg
-      width={width}
-      viewBox="0 0 68 75"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M53.4995 15.3797L53.4995 15.3798C61.5002 23.2119 61.5002 35.9083 53.4995 43.7413L39.7606 57.1923C39.3415 57.6026 38.6585 57.6026 38.2394 57.1923L24.5005 43.7413C16.4999 35.9084 16.4998 23.2131 24.5002 15.3801C24.5064 15.3741 24.5123 15.3679 24.5181 15.3617C32.5262 7.54009 45.499 7.54611 53.4995 15.3797ZM24.6264 14.8705C24.6665 14.9967 24.6584 15.1387 24.5894 15.2627C24.6567 15.1397 24.6657 14.9973 24.6264 14.8705ZM24.1503 15.5225C24.0017 15.5225 23.881 15.4657 23.7955 15.3787C23.7957 15.3784 23.796 15.3781 23.7962 15.3779C23.8828 15.4655 24.0041 15.5225 24.1503 15.5225Z"
-        stroke="#007A49"
-      />
-      <path
-        d="M43.4995 15.3797L43.4995 15.3798C51.5002 23.2119 51.5002 35.9083 43.4995 43.7413L29.7606 57.1923C29.3415 57.6026 28.6585 57.6026 28.2394 57.1923L14.5005 43.7413C6.49994 35.9084 6.49983 23.2131 14.5002 15.3801C14.5064 15.3741 14.5123 15.3679 14.5181 15.3617C22.5262 7.54009 35.499 7.54611 43.4995 15.3797ZM14.6264 14.8705C14.6665 14.9967 14.6584 15.1387 14.5894 15.2627C14.6567 15.1397 14.6657 14.9973 14.6264 14.8705ZM14.1503 15.5225C14.0017 15.5225 13.881 15.4657 13.7955 15.3787C13.7957 15.3784 13.796 15.3781 13.7962 15.3779C13.8828 15.4655 14.0041 15.5225 14.1503 15.5225Z"
-        stroke="#007A49"
-      />
-      <g filter="url(#filter0_d_3057_1324)">
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M19.1503 15.0225C19.1504 15.0225 19.1505 15.0227 19.1504 15.0228C10.9498 23.0519 10.9499 36.0696 19.1507 44.0986L32.8896 57.5495C33.5031 58.1502 34.4969 58.1502 35.1104 57.5495L48.8493 44.0986C57.0502 36.0695 57.0502 23.0507 48.8493 15.0225C40.6484 6.99261 27.3511 6.9925 19.1501 15.0222C19.15 15.0223 19.1501 15.0225 19.1503 15.0225V15.0225Z"
-          fill="white"
-        />
-        <path
-          d="M48.3246 15.5584L48.3246 15.5584C56.2251 23.2925 56.2252 35.8277 48.3246 43.5627L34.5857 57.0136C34.2638 57.3288 33.7362 57.3288 33.4143 57.0136L19.6754 43.5627C11.775 35.8278 11.7749 23.2937 19.6751 15.5587C19.6826 15.5514 19.6899 15.5439 19.697 15.5364C27.6082 7.81388 40.4226 7.8212 48.3246 15.5584ZM18.6223 15.5586C18.6219 15.559 18.6215 15.5594 18.6211 15.5598C18.7492 15.6887 18.9292 15.7725 19.1503 15.7725C18.9327 15.7725 18.7519 15.6883 18.6223 15.5586Z"
-          stroke="#007A49"
-          strokeWidth="1.5"
-        />
-      </g>
-      <path
-        d="M49.5717 24.377C49.154 24.2899 48.7363 24.1919 48.3187 24.1266C47.3334 23.9742 46.4553 24.3661 45.5664 24.6927C44.4312 25.1172 43.4459 25.8139 42.45 26.4997C41.5397 27.1311 40.6401 27.7843 39.6977 28.3503C39.1729 28.666 38.5732 28.8511 37.9949 29.0362C37.5023 29.1886 36.9882 29.1341 36.5063 28.9491C36.3349 28.8837 36.2493 28.9164 36.1636 29.0688C35.3068 30.5057 34.6 31.9971 34.2252 33.6409C34.1074 34.1416 33.829 34.49 33.347 34.6315C32.8651 34.773 32.4153 34.675 32.0726 34.294C30.7768 32.868 29.481 31.431 28.1851 29.9941C27.9817 29.7764 28.0031 29.5478 28.2173 29.3518C28.3565 29.2212 28.4957 29.0906 28.6349 28.9708C29.2561 28.4374 30.07 28.481 30.6162 29.0797C31.2266 29.7546 31.8263 30.4404 32.426 31.1262C32.501 31.2133 32.5438 31.344 32.6509 31.431C34.8035 27.3815 37.9735 24.475 41.6146 22.0257C40.276 22.4502 39.1087 23.2122 38.0377 24.1157C36.9775 25.0193 35.9708 25.9881 34.932 26.9352C34.9106 26.2058 34.9641 25.4547 35.1462 24.7253C36.185 20.5996 40.4795 18.4333 44.4312 20.0227C45.2665 20.3601 46.059 20.7847 46.7123 21.4052C47.7082 22.3413 48.6614 23.332 49.6252 24.3008C49.6038 24.3226 49.5824 24.3443 49.561 24.3661L49.5717 24.377Z"
-        fill="#007A49"
-      />
-      <path
-        d="M41.1435 28.0461C42.2145 31.3663 41.1864 35.9057 37.3524 38.3986C33.6684 40.7935 28.7635 40.1077 25.8506 36.8201C22.8948 33.4782 22.7021 28.5142 25.4115 25.009C28.2067 21.384 32.726 20.8179 35.7353 22.1133C35.4676 22.6141 35.1998 23.104 34.9642 23.5612C34.2467 23.4523 33.5292 23.2672 32.8224 23.2455C28.9349 23.104 25.7221 25.9234 25.2509 29.8532C24.7904 33.685 27.5962 37.408 31.3659 37.9741C35.5104 38.5945 39.2266 35.7642 39.762 31.5296C39.8584 30.8003 39.762 30.0383 39.6871 29.298C39.6656 29.0368 39.6978 28.9061 39.9334 28.7755C40.3403 28.5578 40.7259 28.2965 41.1435 28.0461Z"
-        fill="#007A49"
-      />
-      <defs>
-        <filter
-          id="filter0_d_3057_1324"
-          x="0.291361"
-          y="0.103952"
-          width="67.4173"
-          height="74.4173"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB"
-        >
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dy="3.81259" />
-          <feGaussianBlur stdDeviation="6.35432" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0.0811932 0 0 0 0 0.0811932 0 0 0 0 0.092806 0 0 0 0.15 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="BackgroundImageFix"
-            result="effect1_dropShadow_3057_1324"
-          />
-          <feBlend
-            mode="normal"
-            in="SourceGraphic"
-            in2="effect1_dropShadow_3057_1324"
-            result="shape"
-          />
-        </filter>
-      </defs>
-    </svg>
-  );
-};
-export default RegisteredTreeClusterMarkerIcon;
